Improve x/profiles keeper files organization
See PR #440
